+++ Sources/KeyboardKit/Previews/Actions+Preview.swift
@@ -10,15 +10,11 @@ import CoreGraphics
 
 public extension KeyboardActionHandler where Self == PreviewKeyboardActionHandler {
     
-    /**
-     This preview handler can be used in SwiftUI previews.
-     */
+    /// This handler can be used in SwiftUI previews.
     static var preview: KeyboardActionHandler { PreviewKeyboardActionHandler() }
 }
 
-/**
- This action handler can be used in SwiftUI previews.
- */
+/// This action handler can be used in SwiftUI previews.
 public class PreviewKeyboardActionHandler: KeyboardActionHandler {
     
     public init() {}