+++ Sources/KeyboardKit/_Deprecated/Callouts+Deprecated.swift
@@ -1,5 +1,12 @@
 import Foundation
 
+@available(*, deprecated, renamed: "KeyboardCalloutContext.ActionContext")
+public typealias ActionCalloutContext = KeyboardCalloutContext.ActionContext
+
+@available(*, deprecated, renamed: "KeyboardCalloutContext.InputContext")
+public typealias InputCalloutContext = KeyboardCalloutContext.InputContext
+
+
 @available(*, deprecated, renamed: "KeyboardCalloutStyle")
 public typealias CalloutStyle = KeyboardCalloutStyle